Moved test files into test folder
